requirements.txt
langchain==0.0.274
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.27.7
fastapi==0.88.0
uvicorn==0.18.3
streamlit==1.25.0
tqdm==4.66.1
